feat: 로그인 여부를 확인하는 어노테이션 생성
