+++ src/main/java/com/donation/auth/LoginInfo.java
@@ -1,0 +1,11 @@
+package com.donation.auth;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.PARAMETER)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface LoginInfo {
+}
