+++ src/main/java/com/donation/web/controller/auth/AuthController.java
@@ -1,0 +1,4 @@
+package com.donation.web.controller.auth;
+
+public class AuthController {
+}
